Revert "add sample webapp and resource in the same application"
