+++ aad/azure-spring-boot-sample-active-directory-resource-server-obo/src/main/resources/application.yml
@@ -12,7 +12,7 @@ azure:
         scopes:
           - https://graph.microsoft.com/User.Read
       webapiB:         # When authorization-grant-type is null, on behalf of flow is used by default
-        authorization-grant-type: on_behalf_of
+        authorization-grant-type: on-behalf-of
         scopes:
           - <Web-API-B-app-id-url>/WebApiB.ExampleScope
       webapiC:
