+++ aad/azure-spring-boot-sample-active-directory-resource-server/README.md
@@ -33,21 +33,7 @@ This sample illustrates how to protect a Java web API by restricting access to i
 1. Under **webapiB** application, select **Expose an API** -> **Add a scope**, Use the default Application ID URI, click **Save and continue** button.![Set application id url](docs/image-set-application-id-url.png "Set application id url")
 1. Wait the page refresh finished. Then set the **Scope name** to `WebApiB.ExampleScope`.![Add a scope](docs/image-add-a-scope.png "Add a scope")
 1. Finally, the api exposed in `webapiB`.![Finally, the API exposed in webAPI](docs/image-expose-api.png "Finally, the API exposed in webAPI")
-1. Expose an API by adding `appRoles` , See [Example: Application app role] for more information about app roles setting.
 
-    ```json
-    {
-        "allowedMemberTypes": [
-          "Application"
-        ],
-        "description": "WebApiB ClientCredential Example Scope",
-        "displayName": "WebApiB ClientCredential Example Scope",
-        "id": "d2bec026-b75f-418d-9493-8462f54f25d9",
-        "isEnabled": true, 
-        "value": "WebApiB.ClientCredential.ExampleScope"
-    }
-    ```
-   
 See [Expose scoped permission to web api] for more information about web api.
 
 ## Examples
@@ -78,7 +64,7 @@ We could use Postman to simulate a Web APP to send a request to a Web API.
 1. The `aud` in access token should be the current Web API.
 
 ```http request
-GET /webapiB HTTP/1.1
+GET /file HTTP/1.1
 Authorization: Bearer eyJ0eXAiO ... 0X2tnSQLEANnSPHY0gKcgw
 ```
 ```http request
@@ -100,4 +86,3 @@ Authorization: Bearer eyJ0eXAiO ... 0X2tnSQLEANnSPHY0gKcgw
 [Resource Server]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/aad/azure-spring-boot-sample-active-directory-resource-server
 [Resource Server Obo]: https://github.com/Azure-Samples/azure-spring-boot-samples/blob/main/aad/azure-spring-boot-sample-active-directory-resource-server-obo
 [resource server password credentials]: https://docs.microsoft.com/azure/active-directory/develop/v2-oauth-ropc
-[Example: Application app role]: https://docs.microsoft.com/azure/active-directory/develop/howto-add-app-roles-in-azure-ad-apps#example-application-app-role
