+++ aad/azure-spring-boot-sample-active-directory-resource-server/src/main/java/com/azure/spring/sample/aad/controller/HomeController.java
@@ -28,7 +28,6 @@ public class HomeController {
     }
 
     @GetMapping("/webapiB/clientCredential")
-    @PreAuthorize("hasAuthority('APPROLE_WebApiB.ClientCredential.ExampleScope')")
     @ResponseBody
     public String clientCredential() {
         return "Response from webApiB: clientCredential";
