--- aad/azure-spring-boot-sample-active-directory-webapp-and-resource-server/src/main/java/com/azure/spring/sample/aad/AzureADOAuth2WebAppAndResourceServerSampleApplication.java
@@ -1,15 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.spring.sample.aad;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AzureADOAuth2WebAppAndResourceServerSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AzureADOAuth2WebAppAndResourceServerSampleApplication.class, args);
-    }
-}
