--- aad/azure-spring-boot-sample-active-directory-webapp-and-resource-server/src/main/java/com/azure/spring/sample/aad/controller/webapp/AuthorityController.java
@@ -1,19 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.spring.sample.aad.controller.webapp;
-
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class AuthorityController {
-    @GetMapping("Admin")
-    @ResponseBody
-    @PreAuthorize("hasAuthority('APPROLE_Admin')")
-    public String admin() {
-        return "Admin message";
-    }
-}
