--- aad/azure-spring-boot-sample-active-directory-webapp-and-resource-server/src/main/resources/application.yml
@@ -1,36 +1,0 @@
-server:
-  port: 8084
-
-azure:
-  activedirectory:
-    client-id: <client-id>
-    client-secret: <client-secret>
-    tenant-id: <tenant-id>
-    user-group:
-      allowed-group-names: <group1>,<group2>
-      allowed-group-ids: <group1-id>,<group2-id>
-    post-logout-redirect-uri: http://localhost:8080
-    application-type: web_application_and_resource_server # This is required.
-    app-id-uri: api://<Web-API-C-app-id-url>
-    authorization-clients:
-      arm:
-        authorizationGrantType: authorization_code # This is required.
-        on-demand: true
-        scopes: https://management.core.windows.net/user_impersonation
-      graph:
-        authorizationGrantType: authorization_code # This is required.
-        scopes:
-          - https://graph.microsoft.com/User.Read
-          - https://graph.microsoft.com/Directory.Read.All
-#      webapiA:  # Web Application uses webapiA client to access restricted resources.
-#        authorizationGrantType: authorization_code
-#        scopes:
-#          - api://<Web-API-A-app-id-url>/Obo.WebApiA.ExampleScope
-#      webapiBWithObo: # Resource server uses webapiBObo client to access restricted resources.
-#        authorization-grant-type: on_behalf_of
-#        scopes:
-#          - api:/<Web-API-B-app-id-url>/WebApiB.ExampleScope
-#      webapiBWithClientCredentials:  # Both Web Application or Resource Server can use webapiBWithClientCredentials client to access restricted resources.
-#        authorization-grant-type: client_credentials
-#        scopes:
-#          - api://<Web-API-B-app-id-url>/.default
