--- aad/azure-spring-boot-sample-active-directory-webapp-and-resource-server/src/main/resources/templates/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4">
-<head>
-    <title>Spring Security - OAuth 2.0 Login</title>
-    <meta charset="utf-8" />
-</head>
-<body>
-<div style="float: right" th:fragment="logout" sec:authorize="isAuthenticated()">
-    <div>
-        <span style="font-weight:bold">User: </span><span sec:authentication="name"></span>
-    </div>
-    <div>
-        <span style="font-weight:bold">Default Authorities: </span><span sec:authentication="principal.authorities"></span>
-    </div>
-    <div>
-        <span style="font-weight:bold">Mapped Authorities: </span><span sec:authentication="authorities"></span>
-    </div>
-    <div>&nbsp;</div>
-    <div style="float:right">
-        <form action="#" th:action="@{/logout}" method="post">
-            <input type="submit" value="Logout" />
-        </form>
-    </div>
-</div>
-<h1>Azure Active Directory OAuth 2.0 Login with Spring Security</h1>
-<div>
-    You are successfully logged in <span style="font-weight:bold" th:text="${userName}"></span>
-    (<span style="font-weight:bold" th:text="${preferredUsername}"></span>)
-    via the OAuth 2.0 Client <span style="font-weight:bold" th:text="${clientName}"></span>
-</div>
-<div>&nbsp;</div>
-<div>&nbsp;</div>
-<div>
-    <a href="/group1" >Group1 Message</a> |
-    <a href="/group2" >Group2 Message</a> |
-    <a href="/Admin" >Admin Message</a> |
-    <a href="/graph" >Graph Client</a> |
-    <a href="/arm" >Arm Client</a> |
-    <!--<a href="/webapp/webapiA/webapiB" >WebapiA Obo Client</a> |
-    <a href="/webapp/webapiB/clientCredential" >WebapiB Client Credentials</a>-->
-</div>
-</body>
-</html>
