+++ aad/azure-spring-boot-sample-active-directory-webapp/src/main/resources/templates/index.html
@@ -37,7 +37,7 @@
     <a href="/graph" >Graph Client</a> |
     <a href="/arm" >Arm Client</a> |
     <a href="/webapp/webapiA/webapiB" >Obo Client</a> |
-    <a href="/webapp/webapiB/clientCredential" >Client Credential Client</a>
+    <a href="/webapp/webapiB/clientCredential" >Client Credential Client</a> |
 </div>
 </body>
 </html>
